pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "openai_agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.68.2",
    "openai-agents>=0.0.15",
    "python-dotenv>=1.0.1",
    "sendgrid>=6.11.0",
    "gradio>=5.22.0",
    "pydantic>=2.0.0",
    "ipykernel>=6.29.5",
]